add: Struct

+++ basic/basic.sol
@@ -2,9 +2,9 @@
 pragma solidity >=0.7.0 <0.9.0;
 
 contract basic {
-    uint256 test_num = 6;
+    uint8 test_num = 6;
 
-    function call_num() public view returns (uint256) {
+    function call_num() public view returns (uint8) {
         return test_num;
     }
 }
